Exception Management Refactor v2
