+++ python/api/api_app.py
@@ -21,7 +21,8 @@ from fastapi.responses import JSONResponse
 from python.api.endpoints import bars_router, broker_router
 from python.configuration.app_config_manager import AppConfigManager
 from python.data_management.index.bars_index_manager import BarsIndexManager
-from python.framework.types.api.api_types import ApiException, BrokerListResponse, HealthResponse, TimeframeInfo, TimeframeListResponse
+from python.framework.exceptions.api_errors import ApiException
+from python.framework.types.api.api_types import BrokerListResponse, HealthResponse, TimeframeInfo, TimeframeListResponse
 from python.framework.utils.timeframe_config_utils import TimeframeConfig
