+++ python/framework/types/api/api_types.py
@@ -49,16 +49,3 @@ class TimeframeInfo(BaseModel):
 
 class TimeframeListResponse(BaseModel):
     timeframes: list[TimeframeInfo]
-
-
-class ApiException(Exception):
-    """
-    Raised by endpoint handlers to produce a structured JSON error response.
-
-    Response body: {"error": "<error>", "detail": "<detail>"}
-    """
-
-    def __init__(self, status_code: int, error: str, detail: str):
-        self.status_code = status_code
-        self.error = error
-        self.detail = detail
